refactor(websocket): modularize server and add rollback from history func
