+++ server/websocket/src/bin/main.rs
@@ -1,246 +1,16 @@
-use axum::{
-    extract::{
-        ws::{WebSocket, WebSocketUpgrade},
-        Path,
-    },
-    http::StatusCode,
-    response::{IntoResponse, Response},
-    routing::get,
-    Json, Router,
+use std::sync::Arc;
+
+use tracing::error;
+use websocket::{
+    conf::Config,
+    pool::BroadcastPool,
+    server::{create_router, ensure_bucket, start_server},
+    storage::gcs::GcsStore,
+    AppState,
 };
 
-#[cfg(feature = "auth")]
-use axum::body::Bytes;
-#[cfg(feature = "auth")]
-use axum::extract::Query;
-use google_cloud_storage::{
-    client::Client,
-    http::buckets::insert::{BucketCreationConfig, InsertBucketRequest},
-};
-#[cfg(feature = "auth")]
-use serde::Deserialize;
-use serde::Serialize;
-use std::sync::Arc;
-use time::OffsetDateTime;
 #[cfg(feature = "auth")]
 use websocket::auth::AuthService;
-use websocket::conf::Config;
-use websocket::group::BroadcastGroup;
-use websocket::pool::BroadcastPool;
-use websocket::storage::gcs::GcsStore;
-use websocket::storage::kv::DocOps;
-use websocket::ws::WarpConn;
-use yrs::updates::encoder::Encode;
-use yrs::{Doc, ReadTxn, StateVector, Transact};
-
-const BUCKET_NAME: &str = "yrs-dev";
-const PORT: &str = "8000";
-
-#[cfg(feature = "auth")]
-#[derive(Debug, Deserialize)]
-struct AuthQuery {
-    #[serde(default)]
-    token: String,
-}
-
-#[derive(Serialize)]
-struct UpdateHistoryEntry {
-    update: Vec<u8>,
-    clock: u32,
-    timestamp: OffsetDateTime,
-}
-
-async fn ensure_bucket(client: &Client) -> Result<(), anyhow::Error> {
-    let bucket = BucketCreationConfig {
-        location: "US".to_string(),
-        ..Default::default()
-    };
-    let request = InsertBucketRequest {
-        name: BUCKET_NAME.to_string(),
-        bucket,
-        ..Default::default()
-    };
-
-    match client.insert_bucket(&request).await {
-        Ok(_) => Ok(()),
-        Err(e) if e.to_string().contains("already exists") => Ok(()),
-        Err(e) => Err(e.into()),
-    }
-}
-
-#[cfg(feature = "auth")]
-#[derive(Clone)]
-struct AppState {
-    pool: Arc<BroadcastPool>,
-    auth: Arc<AuthService>,
-}
-
-#[cfg(not(feature = "auth"))]
-#[derive(Clone)]
-struct AppState {
-    pool: Arc<BroadcastPool>,
-}
-
-async fn ws_handler(
-    ws: WebSocketUpgrade,
-    Path(doc_id): Path<String>,
-    #[cfg(feature = "auth")] Query(auth): Query<AuthQuery>,
-    axum::extract::State(state): axum::extract::State<AppState>,
-) -> Response {
-    let doc_id = if doc_id.ends_with(":main") {
-        doc_id[..doc_id.len() - 5].to_string()
-    } else {
-        doc_id.clone()
-    };
-
-    #[cfg(feature = "auth")]
-    {
-        match state.auth.verify_token(&auth.token).await {
-            Ok(true) => (),
-            Ok(false) => {
-                tracing::warn!(
-                    "Authentication failed for doc_id: {}, token: {}",
-                    doc_id,
-                    auth.token
-                );
-                return Bytes::from("Unauthorized").into_response();
-            }
-            Err(e) => {
-                tracing::error!("Authentication error: {}", e);
-                return Bytes::from("Internal Server Error").into_response();
-            }
-        }
-    }
-
-    let bcast = match state.pool.get_or_create_group(&doc_id).await {
-        Ok(group) => group,
-        Err(e) => {
-            tracing::error!("Failed to get or create group for {}: {}", doc_id, e);
-            return Response::builder()
-                .status(500)
-                .body(axum::body::Body::empty())
-                .unwrap();
-        }
-    };
-    ws.on_upgrade(move |socket| handle_socket(socket, bcast, doc_id, state.pool.clone()))
-}
-
-async fn handle_socket(
-    socket: WebSocket,
-    bcast: Arc<BroadcastGroup>,
-    doc_id: String,
-    pool: Arc<BroadcastPool>,
-) {
-    bcast.increment_connections();
-    let conn = WarpConn::new(bcast, socket);
-    if let Err(e) = conn.await {
-        tracing::error!("WebSocket connection error: {}", e);
-    }
-    pool.remove_connection(&doc_id).await;
-}
-
-async fn get_latest_doc(
-    Path(doc_id): Path<String>,
-    #[cfg(feature = "auth")] Query(auth): Query<AuthQuery>,
-    axum::extract::State(state): axum::extract::State<AppState>,
-) -> impl IntoResponse {
-    #[cfg(feature = "auth")]
-    {
-        match state.auth.verify_token(&auth.token).await {
-            Ok(true) => (),
-            Ok(false) => {
-                tracing::warn!(
-                    "Authentication failed for doc_id: {}, token: {}",
-                    doc_id,
-                    auth.token
-                );
-                return (StatusCode::UNAUTHORIZED, "Unauthorized").into_response();
-            }
-            Err(e) => {
-                tracing::error!("Authentication error: {}", e);
-                return (StatusCode::INTERNAL_SERVER_ERROR, "Internal Server Error")
-                    .into_response();
-            }
-        }
-    }
-
-    let doc_id = if doc_id.ends_with(":main") {
-        doc_id[..doc_id.len() - 5].to_string()
-    } else {
-        doc_id.clone()
-    };
-
-    let store = state.pool.get_store();
-    let doc = Doc::new();
-    let mut txn = doc.transact_mut();
-
-    match store.load_doc(&doc_id, &mut txn).await {
-        Ok(true) => {
-            drop(txn);
-            let read_txn = doc.transact();
-            let state = read_txn.encode_diff_v1(&StateVector::default());
-            (StatusCode::OK, Json(state)).into_response()
-        }
-        Ok(false) => (StatusCode::NOT_FOUND, "Document not found").into_response(),
-        Err(e) => {
-            tracing::error!("Failed to get document: {}", e);
-            (StatusCode::INTERNAL_SERVER_ERROR, "Internal Server Error").into_response()
-        }
-    }
-}
-
-async fn get_doc_history(
-    Path(doc_id): Path<String>,
-    #[cfg(feature = "auth")] Query(auth): Query<AuthQuery>,
-    axum::extract::State(state): axum::extract::State<AppState>,
-) -> impl IntoResponse {
-    #[cfg(feature = "auth")]
-    {
-        match state.auth.verify_token(&auth.token).await {
-            Ok(true) => (),
-            Ok(false) => {
-                tracing::warn!(
-                    "Authentication failed for doc_id: {}, token: {}",
-                    doc_id,
-                    auth.token
-                );
-                return (StatusCode::UNAUTHORIZED, "Unauthorized").into_response();
-            }
-            Err(e) => {
-                tracing::error!("Authentication error: {}", e);
-                return (StatusCode::INTERNAL_SERVER_ERROR, "Internal Server Error")
-                    .into_response();
-            }
-        }
-    }
-
-    let doc_id = if doc_id.ends_with(":main") {
-        doc_id[..doc_id.len() - 5].to_string()
-    } else {
-        doc_id.clone()
-    };
-
-    let store = state.pool.get_store();
-
-    // Get all updates for this document
-    match store.get_updates(&doc_id).await {
-        Ok(updates) => {
-            let history: Vec<UpdateHistoryEntry> = updates
-                .into_iter()
-                .map(|info| UpdateHistoryEntry {
-                    update: info.update.encode_v1(),
-                    clock: info.clock,
-                    timestamp: info.timestamp,
-                })
-                .collect();
-            (StatusCode::OK, Json(history)).into_response()
-        }
-        Err(e) => {
-            tracing::error!("Failed to get document history: {}", e);
-            (StatusCode::INTERNAL_SERVER_ERROR, "Internal Server Error").into_response()
-        }
-    }
-}
 
 #[tokio::main]
 async fn main() {
@@ -254,7 +24,7 @@ async fn main() {
     let config = match Config::load() {
         Ok(config) => config,
         Err(e) => {
-            tracing::error!("Failed to load config: {}", e);
+            error!("Failed to load config: {}", e);
             std::process::exit(1);
         }
     };
@@ -265,7 +35,7 @@ async fn main() {
 
     // Ensure bucket exists
     if let Err(e) = ensure_bucket(&store.client).await {
-        tracing::error!("Failed to ensure bucket exists: {}", e);
+        error!("Failed to ensure bucket exists: {}", e);
         std::process::exit(1);
     }
 
@@ -289,59 +59,10 @@ async fn main() {
         }
     };
 
-    let app = Router::new()
-        .route("/{doc_id}", get(ws_handler))
-        .route("/{doc_id}/latest", get(get_latest_doc))
-        .route("/{doc_id}/history", get(get_doc_history))
-        .with_state(state);
+    let app = create_router(state);
 
-    tracing::info!("Starting server on 0.0.0.0:{}", PORT);
-    let listener = tokio::net::TcpListener::bind(format!("0.0.0.0:{}", PORT))
-        .await
-        .map_err(|e| {
-            tracing::error!("Failed to bind to port {}: {}", PORT, e);
-            std::process::exit(1);
-        })
-        .unwrap();
-
-    // Setup shutdown signal handler
-    let (tx, _) = tokio::sync::broadcast::channel(1);
-    let shutdown_signal = tx.clone();
-
-    tokio::spawn(async move {
-        let ctrl_c = async {
-            tokio::signal::ctrl_c()
-                .await
-                .expect("Failed to install Ctrl+C handler");
-        };
-
-        #[cfg(unix)]
-        let terminate = async {
-            tokio::signal::unix::signal(tokio::signal::unix::SignalKind::terminate())
-                .expect("Failed to install signal handler")
-                .recv()
-                .await;
-        };
-
-        #[cfg(not(unix))]
-        let terminate = std::future::pending::<()>();
-
-        tokio::select! {
-            _ = ctrl_c => {},
-            _ = terminate => {},
-        }
-
-        tracing::info!("Shutdown signal received");
-        let _ = shutdown_signal.send(());
-    });
-
-    if let Err(e) = axum::serve(listener, app)
-        .with_graceful_shutdown(async move {
-            let _ = tx.subscribe().recv().await;
-        })
-        .await
-    {
-        tracing::error!("Server error: {}", e);
+    if let Err(e) = start_server(app).await {
+        error!("Server error: {}", e);
         std::process::exit(1);
     }
 }
